chore: configstack code refactoring
