+++ cli/commands/graph-dependencies/action.go
@@ -9,7 +9,7 @@ import (
 
 // Run graph dependencies prints the dependency graph to stdout
 func Run(ctx context.Context, opts *options.TerragruntOptions) error {
-	stack, err := configstack.FindStackInSubfolders(ctx, opts, nil)
+	stack, err := configstack.FindStackInSubfolders(ctx, opts)
 	if err != nil {
 		return err
 	}
