+++ cli/commands/output-module-groups/action.go
@@ -9,7 +9,7 @@ import (
 )
 
 func Run(ctx context.Context, opts *options.TerragruntOptions) error {
-	stack, err := configstack.FindStackInSubfolders(ctx, opts, nil)
+	stack, err := configstack.FindStackInSubfolders(ctx, opts)
 	if err != nil {
 		return err
 	}
