+++ configstack/errors.go
@@ -1,6 +1,11 @@
 package configstack
 
-import "fmt"
+import (
+	"fmt"
+	"strings"
+
+	"github.com/gruntwork-io/terragrunt/shell"
+)
 
 // Custom error types
 
@@ -31,4 +36,38 @@ type InfiniteRecursion struct {
 
 func (err InfiniteRecursion) Error() string {
 	return fmt.Sprintf("Hit what seems to be an infinite recursion after going %d levels deep. Please check for a circular dependency! Modules involved: %v", err.RecursionLevel, err.Modules)
+}
+
+var NoTerraformModulesFound = fmt.Errorf("Could not find any subfolders with Terragrunt configuration files")
+
+type DependencyCycle []string
+
+func (err DependencyCycle) Error() string {
+	return fmt.Sprintf("Found a dependency cycle between modules: %s", strings.Join([]string(err), " -> "))
+}
+
+type DependencyFinishedWithError struct {
+	Module     *TerraformModule
+	Dependency *TerraformModule
+	Err        error
+}
+
+func (err DependencyFinishedWithError) Error() string {
+	return fmt.Sprintf("Cannot process module %s because one of its dependencies, %s, finished with an error: %s", err.Module, err.Dependency, err.Err)
+}
+
+func (this DependencyFinishedWithError) ExitStatus() (int, error) {
+	if exitCode, err := shell.GetExitCode(this.Err); err == nil {
+		return exitCode, nil
+	}
+	return -1, this
+}
+
+type DependencyNotFoundWhileCrossLinking struct {
+	Module     *runningModule
+	Dependency *TerraformModule
+}
+
+func (err DependencyNotFoundWhileCrossLinking) Error() string {
+	return fmt.Sprintf("Module %v specifies a dependency on module %v, but could not find that module while cross-linking dependencies. This is most likely a bug in Terragrunt. Please report it.", err.Module, err.Dependency)
 }
